requirement.txt
beautifulsoup4==4.8.1
bs4==0.0.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Cython==0.29.14
google-api-core==1.14.3
google-auth==1.7.0
google-cloud-language==1.3.0
google-images-download==2.8.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
idna==2.8
joblib==0.14.0
JPype1==0.6.3
konlpy==0.5.1
mecab-python===0.996-ko-0.9.2-msvc
nltk==3.4.5
numpy==1.17.3
pandas==0.25.3
Pillow==6.2.1
protobuf==3.10.0
psutil==5.6.3
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
selenium==3.141.0
six==1.12.0
soupsieve==1.9.4
soynlp==0.0.493
urllib3==1.25.6
